limit usage of ccxt to freqtrade/exchange only
