+++ freqtrade/data/converter.py
@@ -2,9 +2,9 @@
 Functions to convert data from one format to another
 """
 import logging
+
 import pandas as pd
 from pandas import DataFrame, to_datetime
-from freqtrade.misc import timeframe_to_minutes
 
 
 logger = logging.getLogger(__name__)
@@ -58,6 +58,8 @@ def ohlcv_fill_up_missing_data(dataframe: DataFrame, ticker_interval: str) -> Da
     using the previous close as price for "open", "high" "low" and "close", volume is set to 0
 
     """
+    from freqtrade.exchange import timeframe_to_minutes
+
     ohlc_dict = {
         'open': 'first',
         'high': 'max',
