+++ freqtrade/exchange/__init__.py
@@ -1,3 +1,8 @@
 from freqtrade.exchange.exchange import Exchange  # noqa: F401
+from freqtrade.exchange.exchange import (is_exchange_supported,  # noqa: F401
+                                         supported_exchanges)
+from freqtrade.exchange.exchange import (timeframe_to_seconds,  # noqa: F401
+                                         timeframe_to_minutes,
+                                         timeframe_to_msecs)
 from freqtrade.exchange.kraken import Kraken  # noqa: F401
 from freqtrade.exchange.binance import Binance  # noqa: F401
